refactor: consolidate isFleximarkWorkspace functionality into fs module
